test(lsp): adds functionality for testing language server
