+++ crates/mun_language_server/src/lib.rs
@@ -1,3 +1,13 @@
+use std::convert::TryFrom;
+
+use serde::{de::DeserializeOwned, Serialize};
+
+pub use config::{Config, FilesWatcher};
+pub use main_loop::main_loop;
+use paths::AbsPathBuf;
+use project::ProjectManifest;
+pub(crate) use state::LanguageServerState;
+
 mod analysis;
 mod cancelation;
 mod capabilities;
@@ -6,34 +16,25 @@ mod config;
 mod conversion;
 mod db;
 mod diagnostics;
-mod dispatcher;
 mod main_loop;
-mod project_manifest;
-mod workspace;
-
-pub use config::Config;
-pub use main_loop::main_loop;
-
-use crate::{config::FilesWatcher, project_manifest::ProjectManifest};
-use paths::AbsPathBuf;
-use serde::{de::DeserializeOwned, Serialize};
-use std::convert::TryFrom;
-
-pub type Result<T> = anyhow::Result<T>;
+mod state;
 
 /// Deserializes a `T` from a json value.
-pub fn from_json<T: DeserializeOwned>(what: &'static str, json: serde_json::Value) -> Result<T> {
+pub fn from_json<T: DeserializeOwned>(
+    what: &'static str,
+    json: serde_json::Value,
+) -> anyhow::Result<T> {
     T::deserialize(&json)
         .map_err(|e| anyhow::anyhow!("could not deserialize {}: {}: {}", what, e, json))
 }
 
 /// Converts the `T` to a json value
-pub fn to_json<T: Serialize>(value: T) -> Result<serde_json::Value> {
+pub fn to_json<T: Serialize>(value: T) -> anyhow::Result<serde_json::Value> {
     serde_json::to_value(value).map_err(|e| anyhow::anyhow!("could not serialize to json: {}", e))
 }
 
 /// Main entry point for the language server
-pub fn run_server() -> Result<()> {
+pub fn run_server() -> anyhow::Result<()> {
     log::info!("language server started");
 
     // Setup IO connections
