+++ crates/mun_language_server/src/state/utils.rs
@@ -1,0 +1,74 @@
+use super::LanguageServerState;
+
+#[derive(Debug, Eq, PartialEq)]
+pub(crate) enum Progress {
+    Begin,
+    Report,
+    End,
+}
+
+impl Progress {
+    /// Builds a fractional progress value
+    pub(crate) fn fraction(done: usize, total: usize) -> f64 {
+        assert!(done <= total);
+        done as f64 / total.max(1) as f64
+    }
+}
+
+impl LanguageServerState {
+    /// Sends a notification to the client to display the specified message to the user.
+    pub(crate) fn show_message(&mut self, typ: lsp_types::MessageType, message: impl AsRef<str>) {
+        let message = message.as_ref().to_owned();
+        self.send_notification::<lsp_types::notification::ShowMessage>(
+            lsp_types::ShowMessageParams { typ, message },
+        )
+    }
+
+    /// Reports progress to the user via the `WorkDoneProgress` protocol.
+    pub(crate) fn report_progress(
+        &mut self,
+        title: &str,
+        state: Progress,
+        message: Option<String>,
+        fraction: Option<f64>,
+    ) {
+        // TODO: Ensure that the client supports WorkDoneProgress
+
+        let percentage = fraction.map(|f| {
+            (0.0..=1.0).contains(&f);
+            (f * 100.0) as u32
+        });
+        let token = lsp_types::ProgressToken::String(format!("mun/{}", title));
+        let work_done_progress = match state {
+            Progress::Begin => {
+                self.send_request::<lsp_types::request::WorkDoneProgressCreate>(
+                    lsp_types::WorkDoneProgressCreateParams {
+                        token: token.clone(),
+                    },
+                    |_, _| (),
+                );
+
+                lsp_types::WorkDoneProgress::Begin(lsp_types::WorkDoneProgressBegin {
+                    title: title.into(),
+                    cancellable: None,
+                    message,
+                    percentage,
+                })
+            }
+            Progress::Report => {
+                lsp_types::WorkDoneProgress::Report(lsp_types::WorkDoneProgressReport {
+                    cancellable: None,
+                    message,
+                    percentage,
+                })
+            }
+            Progress::End => {
+                lsp_types::WorkDoneProgress::End(lsp_types::WorkDoneProgressEnd { message })
+            }
+        };
+        self.send_notification::<lsp_types::notification::Progress>(lsp_types::ProgressParams {
+            token,
+            value: lsp_types::ProgressParamsValue::WorkDone(work_done_progress),
+        });
+    }
+}
