+++ crates/mun_language_server/tests/support.rs
@@ -1,43 +1,131 @@
 use crossbeam_channel::{after, select};
 use lsp_server::{Connection, Message, Notification, Request};
-use lsp_types::{notification::Exit, request::Shutdown};
-use mun_language_server::{main_loop, Config};
+use lsp_types::{
+    notification::Exit, request::Shutdown, ProgressParams, ProgressParamsValue, WorkDoneProgress,
+};
+use mun_language_server::{main_loop, Config, FilesWatcher};
+use mun_test::Fixture;
 use paths::AbsPathBuf;
+use project::ProjectManifest;
 use serde::Serialize;
 use serde_json::Value;
-use std::convert::TryFrom;
-use std::time::Duration;
+use std::{
+    cell::{Cell, RefCell},
+    convert::TryInto,
+    fs,
+    time::Duration,
+};
+
+/// A `Project` represents a project that a language server can work with. Call the [`server`]
+/// method to instantiate a language server that will serve information about the project.
+pub struct Project<'a> {
+    fixture: &'a str,
+    tmp_dir: Option<tempdir::TempDir>,
+}
+
+impl<'a> Project<'a> {
+    /// Constructs a project from a fixture.
+    pub fn with_fixture(fixture: &str) -> Project {
+        Project {
+            fixture,
+            tmp_dir: None,
+        }
+    }
+
+    /// Instantiates a language server for this project.
+    pub fn server(self) -> Server {
+        // Get or create a temporary directory
+        let tmp_dir = self
+            .tmp_dir
+            .unwrap_or_else(|| tempdir::TempDir::new("testdir").unwrap());
+
+        // Write all fixtures to a folder
+        for entry in Fixture::parse(self.fixture) {
+            let path = entry.relative_path.to_path(tmp_dir.path());
+            fs::create_dir_all(path.parent().unwrap()).unwrap();
+            fs::write(path.as_path(), entry.text.as_bytes()).unwrap();
+        }
+
+        let tmp_dir_path: AbsPathBuf = tmp_dir
+            .path()
+            .to_path_buf()
+            .try_into()
+            .expect("could not convert temp dir to absolute path");
+        let roots = vec![tmp_dir_path.clone()];
+
+        let discovered_projects = ProjectManifest::discover_all(roots.into_iter());
+
+        // Construct a default configuration for the server
+        let config = Config {
+            discovered_projects: Some(discovered_projects),
+            watcher: FilesWatcher::Client,
+            ..Config::new(tmp_dir_path)
+        };
+
+        // TODO: Provide the ability to modify the configuration externally
+
+        Server::new(tmp_dir, config)
+    }
+}
 
 /// An object that runs the language server main loop and enables sending and receiving messages
 /// to and from it.
 pub struct Server {
-    next_request_id: i32,
+    next_request_id: Cell<i32>,
+    messages: RefCell<Vec<Message>>,
     worker: Option<std::thread::JoinHandle<()>>,
     client: Connection,
-    _temp_path: tempdir::TempDir,
+    _tmp_dir: tempdir::TempDir,
 }
 
 impl Server {
     /// Constructs and initializes a new `Server`
-    pub fn new() -> Self {
+    pub fn new(tmp_dir: tempdir::TempDir, config: Config) -> Self {
         let (connection, client) = Connection::memory();
 
-        let temp_path = tempdir::TempDir::new("mun_language_server")
-            .expect("unable to create temporary directory");
-
-        let config = Config::new(
-            AbsPathBuf::try_from(temp_path.path().to_path_buf())
-                .expect("temp_path is not an absolute path"),
-        );
         let worker = std::thread::spawn(move || {
             main_loop(connection, config).unwrap();
         });
 
         Self {
-            next_request_id: Default::default(),
+            next_request_id: Cell::new(1),
+            messages: RefCell::new(Vec::new()),
             worker: Some(worker),
             client,
-            _temp_path: temp_path,
+            _tmp_dir: tmp_dir,
+        }
+    }
+
+    /// Waits until all projects in the workspace have been loaded
+    pub fn wait_until_workspace_is_loaded(self) -> Server {
+        self.wait_for_message_cond(1, &|msg: &Message| match msg {
+            Message::Notification(n) if n.method == "$/progress" => {
+                match n.clone().extract::<ProgressParams>("$/progress").unwrap() {
+                    ProgressParams {
+                        token: lsp_types::ProgressToken::String(ref token),
+                        value: ProgressParamsValue::WorkDone(WorkDoneProgress::End(_)),
+                    } if token == "mun/projects scanned" => true,
+                    _ => false,
+                }
+            }
+            _ => false,
+        });
+        self
+    }
+
+    /// A function to wait for a specific message to arrive
+    fn wait_for_message_cond(&self, n: usize, cond: &dyn Fn(&Message) -> bool) {
+        let mut total = 0;
+        for msg in self.messages.borrow().iter() {
+            if cond(msg) {
+                total += 1
+            }
+        }
+        while total < n {
+            let msg = self.recv().expect("no response");
+            if cond(&msg) {
+                total += 1;
+            }
         }
     }
 
@@ -54,19 +142,19 @@ impl Server {
     }
 
     /// Sends a request to main loop, returning the response
-    fn send_request<R: lsp_types::request::Request>(&mut self, params: R::Params) -> Value
+    fn send_request<R: lsp_types::request::Request>(&self, params: R::Params) -> Value
     where
         R::Params: Serialize,
     {
-        let id = self.next_request_id;
-        self.next_request_id += 1;
+        let id = self.next_request_id.get();
+        self.next_request_id.set(id.wrapping_add(1));
 
         let r = Request::new(id.into(), R::METHOD.to_string(), params);
         self.send_and_receive(r)
     }
 
     /// Sends an LSP notification to the main loop.
-    fn notification<N: lsp_types::notification::Notification>(&mut self, params: N::Params)
+    fn notification<N: lsp_types::notification::Notification>(&self, params: N::Params)
     where
         N::Params: Serialize,
     {
@@ -75,12 +163,12 @@ impl Server {
     }
 
     /// Sends a server notification to the main loop
-    fn send_notification(&mut self, not: Notification) {
+    fn send_notification(&self, not: Notification) {
         self.client.sender.send(Message::Notification(not)).unwrap();
     }
 
     /// Sends a request to the main loop and receives its response
-    fn send_and_receive(&mut self, r: Request) -> Value {
+    fn send_and_receive(&self, r: Request) -> Value {
         let id = r.id.clone();
         self.client.sender.send(r.into()).unwrap();
         while let Some(msg) = self.recv() {
@@ -106,12 +194,16 @@ impl Server {
     }
 
     /// Receives a message from the message or timeout.
-    fn recv(&mut self) -> Option<Message> {
+    fn recv(&self) -> Option<Message> {
         let timeout = Duration::from_secs(120);
-        select! {
+        let msg = select! {
             recv(self.client.receiver) -> msg => msg.ok(),
             recv(after(timeout)) -> _ => panic!("timed out"),
+        };
+        if let Some(ref msg) = msg {
+            self.messages.borrow_mut().push(msg.clone());
         }
+        msg
     }
 }
