+++ crates/mun_project/Cargo.toml
@@ -5,8 +5,10 @@ authors = ["The Mun Team <team@mun-lang.org>"]
 edition = "2018"
 
 [dependencies]
+rustc-hash = "1.1.0"
 serde = "1.0"
 serde_derive = "1.0"
 toml = "0.5"
 semver = { version = "0.10", features = ["serde"] }
 anyhow = "1.0"
+paths = { path="../mun_paths", package="mun_paths" }
