+++ crates/mun_project/src/lib.rs
@@ -1,7 +1,9 @@
-mod manifest;
-mod package;
-
 pub use manifest::{Manifest, ManifestMetadata, PackageId};
 pub use package::Package;
+pub use project_manifest::ProjectManifest;
+
+mod manifest;
+mod package;
+mod project_manifest;
 
 pub const MANIFEST_FILENAME: &str = "mun.toml";
